Create VIOPE_task
The files "mata.txt" and "matb.txt" contain integer matrices of size 10 x 10. Write a program that calculates the sum of the matrices in a new matrix. The resulting sum matrix shall be saved to the file "sum.usr". 

The matrix elements are separated by spaces and newlines. For example: 

1 2 3 4 5 6 7 8 9 10 
11 12 13 14 15 16 17 18 19 20 
... ... ... 
... ... ... 
91 92 93 94 95 96 97 98 99 100 
The sum matrix must be saved in the same format. At the end, the program prints a message indicating successful completion.

Example output:
The sum of the matrices has been calculated into the file sum.usr.
The verification of program output does not account for whitespace characters like "\n", "\t" and " "

task by Virve Prami